fix: use selected content's dateEdited on conflict resolve
this fixes an edge case where the conflicted note was further
updated after the conflict appeared in the front end.
Since we stop syncing if there's a conflict, there's no
way to update the conflicted note.
Instead, we use the selected content's date edited.
This makes sure that when we sync and if there's newer content,
a conflict happens again.