web: show the first existing notebook on note item
previously we were using the notebook at 0 index which failed
in cases where notebook is only moved to trash i.e. its references
from the note item is not removed. If such a note is added to
another notebook it would show nothing because the notebook
at 0 index doesn't exist. This change fixes this issue.